app/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1         
flask-migrate==4.0.5             
psycopg2-binary==2.9.9
python-dotenv==0.21.0        
celery==5.3.6                     
redis==4.5.5                      
flask-jwt-extended==4.6.0
flask-caching==2.1.0
alembic==1.12.1                  
gunicorn==20.1.0
flask-limiter==3.4.0             
python-dateutil==2.8.2
flask-cors==3.0.10
click==8.1.8                      
